+++ libs/game/pxt.json
@@ -3,6 +3,7 @@
     "description": "The game and sprite library - beta",
     "files": [
         "animation.ts",
+        "controller.ts",
         "sprites.ts",
         "sprite.ts",
         "spritemap.ts",
